A2/flask-app/requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0